refactor(table): storybook refactor with TableRowItem
Refactored table storybook which has a separate TableRowItem and handles
additional component options.

NOTE: The 'repeat' directive had to be replaced with 'map' because it
was triggering an error "Storybook preview hooks can only be called
inside decorators and story functions" when using the Checkbox
component, due to use of updateArgs in Checkbox, and existing Storybook
issue # 12006.